feat: initial support of wasm memory64 proposal
* loader
  * limits encoding changes
  * `MemType` missing index type setup
  * assuming default is 32 mode
  * changes related values to uint64_t
* validator
  * introduce helper `getItValueType` to do index type checking (index
    type is a new concept that introduced in the proposal, stands for
    indexing type for memory modes, memory64 use I64 and default is I32)
  * type check
    * store/load instructions
    * atomic instructions
    * v128 instructions
  * page count is possible up to 64-bit now
  * offset is 64bit now
* execution
  * share more code for 32/64-bit mode
  * update related memory instructions
    * instruction load/store and load/store many
    * instruction atomic
* driver: add option
* stack manager: move `popIndexType` that pop value based on index type
  (refers to above description)
* spec test
    * use `wasm-dev` branch to enable memory64 test suites
    * fix `wasmedgeExecutorCoreTests` by fixing the `grow` instruction
    * fix memory limit test (32-bit mode testing)

Later fixup

* fix bus error, overflow problem

  The root cause is combined from two trouble

  1. In executor check we only check the offset didn't overflow by using
     maximal of uint64_t to substract it
  2. In memory check we assume
     `const uint64_t AccessLen = Offset + Length`
     will not overflow!

  but the first check didn't ensure the second won't happen, thus, the
  error occurs.

* fix `f32.load`

  If `ValVariant` stores `uint32_t` then must use `uint32_t`, but I use
  `uint64_t` and cause the problem. `valToIndex` now will judge index type
  of memory64 mode, then use the correct type to expand `ValVariant`.

* fix windows build

+++ include/runtime/instance/module.h
@@ -373,7 +373,7 @@ protected:
     }
     return unsafeGetMemory(Idx);
   }
-  MemoryInstance *unsafeGetMemory(uint32_t Idx) const noexcept {
+  MemoryInstance *unsafeGetMemory(uint64_t Idx) const noexcept {
     return MemInsts[Idx];
   }
   TagInstance *unsafeGetTag(uint32_t Idx) const noexcept {
@@ -409,7 +409,7 @@ protected:
     }
     return unsafeGetData(Idx);
   }
-  DataInstance *unsafeGetData(uint32_t Idx) const noexcept {
+  DataInstance *unsafeGetData(uint64_t Idx) const noexcept {
     return DataInsts[Idx];
   }
